OTA binary files header/lzss handling (#146)
* Initial support for OTA binary encoding

* Code refactoring

* Added no-header feature in binary upload

* Added header decoder

* Added header verification before file upload

* Added header decoder

* removed not required fields in print

* Added unit tests

* dependency update

* rever go upgrade

* Revert "dependency update"

This reverts commit 1256f1f8e4ccac095757a5a00e9775fe9299d040.

* Added support for more boards

* better missing header check

* Removed debug log

* Upgraded go version

* Changed description

* Code clean

* Code updates to include more header info

* Added licenses

* Better ouput formatting based on format output

* Fixed gomod

* Revert "Fixed gomod"

This reverts commit 52b46ce2a1e8ac732e77b98809d5e211be1ed37d.

* Removing not needed license

* Fixed typo in file name

+++ command/ota/massupload.go
@@ -21,12 +21,12 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"io/ioutil"
 	"os"
 	"path/filepath"
 
 	"github.com/arduino/arduino-cloud-cli/config"
 	"github.com/arduino/arduino-cloud-cli/internal/iot"
+	"github.com/arduino/arduino-cloud-cli/internal/ota"
 	iotclient "github.com/arduino/iot-client-go"
 )
 
@@ -37,11 +37,12 @@ const (
 // MassUploadParams contains the parameters needed to
 // perform a Mass OTA upload.
 type MassUploadParams struct {
-	DeviceIDs []string
-	Tags      map[string]string
-	File      string
-	Deferred  bool
-	FQBN      string
+	DeviceIDs        []string
+	Tags             map[string]string
+	File             string
+	Deferred         bool
+	DoNotApplyHeader bool
+	FQBN             string
 }
 
 // Result of an ota upload on a device.
@@ -59,17 +60,35 @@ func MassUpload(ctx context.Context, params *MassUploadParams, cred *config.Cred
 		return nil, errors.New("cannot use both DeviceIDs and Tags. only one of them should be not nil")
 	}
 
-	// Generate .ota file
-	otaDir, err := ioutil.TempDir("", "")
+	_, err := os.Stat(params.File)
 	if err != nil {
-		return nil, fmt.Errorf("%s: %w", "cannot create temporary folder", err)
+		return nil, fmt.Errorf("file %s does not exists: %w", params.File, err)
 	}
-	otaFile := filepath.Join(otaDir, "temp.ota")
-	defer os.RemoveAll(otaDir)
 
-	err = Generate(params.File, otaFile, params.FQBN)
-	if err != nil {
-		return nil, fmt.Errorf("%s: %w", "cannot generate .ota file", err)
+	if !params.DoNotApplyHeader {
+		//Verify if file has already an OTA header
+		header, _ := ota.DecodeOtaFirmwareHeaderFromFile(params.File)
+		if header != nil {
+			params.DoNotApplyHeader = true
+		}
+	}
+
+	// Generate .ota file
+	var otaFile string
+	if params.DoNotApplyHeader {
+		otaFile = params.File
+	} else {
+		otaDir, err := os.MkdirTemp("", "")
+		if err != nil {
+			return nil, fmt.Errorf("%s: %w", "cannot create temporary folder", err)
+		}
+		otaFile := filepath.Join(otaDir, "temp.ota")
+		defer os.RemoveAll(otaDir)
+
+		err = Generate(params.File, otaFile, params.FQBN)
+		if err != nil {
+			return nil, fmt.Errorf("%s: %w", "cannot generate .ota file", err)
+		}
 	}
 
 	iotClient, err := iot.NewClient(cred)
